requirements.txt
aiogram==3.4.1
aiohttp==3.9.3
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.1
PyJWT[crypto]==2.8.0